+++ src/Fable.Transforms/Dart/Replacements.fs
@@ -3450,6 +3450,12 @@ let guids
 
     match i.CompiledName with
     | "NewGuid" -> Helper.LibCall(com, "Guid", "newGuid", t, []) |> Some
+    | "CreateVersion7" ->
+        match args with
+        | [] -> Helper.LibCall(com, "Guid", "createVersion7", t, []) |> Some
+        | _ ->
+            Helper.LibCall(com, "Guid", "createVersion7", t, args, i.SignatureArgTypes)
+            |> Some
     | "Parse" ->
         match args with
         | [ StringConst literalGuid ] -> parseGuid literalGuid