+++ src/Fable.Transforms/Python/Replacements.fs
@@ -3478,6 +3478,12 @@ let guids
     =
     match i.CompiledName with
     | "NewGuid" -> Helper.LibCall(com, "guid", "new_guid", t, []) |> Some
+    | "CreateVersion7" ->
+        match args with
+        | [] -> Helper.LibCall(com, "guid", "create_version7", t, []) |> Some
+        | _ ->
+            Helper.LibCall(com, "guid", "create_version7", t, args, i.SignatureArgTypes)
+            |> Some
     | "Parse" -> Helper.LibCall(com, "guid", "parse", t, args, i.SignatureArgTypes) |> Some
     | "TryParse" -> Helper.LibCall(com, "guid", "try_parse", t, args, i.SignatureArgTypes) |> Some
     | "ToByteArray" ->